+++ malachite-base/src/bin_util/demo_and_bench/slices/latex.rs
@@ -1,0 +1,22 @@
+// Copyright © 2026 Mikhail Hogrefe
+//
+// This file is part of Malachite.
+//
+// Malachite is free software: you can redistribute it and/or modify it under the terms of the GNU
+// Lesser General Public License (LGPL) as published by the Free Software Foundation; either version
+// 3 of the License, or (at your option) any later version. See <https://www.gnu.org/licenses/>.
+
+use malachite_base::strings::latex::ToLatex;
+use malachite_base::test_util::generators::common::{GenConfig, GenMode};
+use malachite_base::test_util::generators::unsigned_vec_gen;
+use malachite_base::test_util::runner::Runner;
+
+pub(crate) fn register(runner: &mut Runner) {
+    register_demo!(runner, demo_slice_to_latex);
+}
+
+fn demo_slice_to_latex(gm: GenMode, config: &GenConfig, limit: usize) {
+    for xs in unsigned_vec_gen::<u8>().get(gm, config).take(limit) {
+        println!("{:?}.to_latex() = {}", xs, xs.as_slice().to_latex());
+    }
+}