+++ malachite-base/src/bin_util/demo_and_bench/slices/mod.rs
@@ -9,6 +9,7 @@
 use malachite_base::test_util::runner::Runner;
 
 pub(crate) fn register(runner: &mut Runner) {
+    latex::register(runner);
     min_repeating_len::register(runner);
     slice_leading_zeros::register(runner);
     slice_move_left::register(runner);
@@ -16,8 +17,10 @@ pub(crate) fn register(runner: &mut Runner) {
     slice_test_zero::register(runner);
     slice_trailing_zeros::register(runner);
     split_into_chunks::register(runner);
+    typst::register(runner);
 }
 
+mod latex;
 mod min_repeating_len;
 mod slice_leading_zeros;
 mod slice_move_left;
@@ -25,3 +28,4 @@ mod slice_set_zero;
 mod slice_test_zero;
 mod slice_trailing_zeros;
 mod split_into_chunks;
+mod typst;